renamed hcablobstore so it isn’t during unittest
