+++ tests/test_gshcablobstore.py
@@ -12,7 +12,7 @@ sys.path.insert(0, pkg_root)  # noqa
 
 from dss.hcablobstore.gs import GSHCABlobStore
 from tests import infra
-from tests.test_hcablobstore import HCABlobStoreTests
+from tests.hcablobstore_base import HCABlobStoreTests
 
 
 class TestGSHCABlobStore(unittest.TestCase, HCABlobStoreTests):
