+++ tests/test_s3hcablobstore.py
@@ -12,7 +12,7 @@ sys.path.insert(0, pkg_root)  # noqa
 
 from dss.hcablobstore.s3 import S3HCABlobStore
 from tests import infra
-from tests.test_hcablobstore import HCABlobStoreTests
+from tests.hcablobstore_base import HCABlobStoreTests
 
 
 class TestS3HCABlobStore(unittest.TestCase, HCABlobStoreTests):
